+++ core/docs/changelog.md
@@ -1,5 +1,7 @@
 # TBD
 
+# 1.58.2
+
 ### Features
 * Container engine server exposes a `RenderTemplatesToFilesArtifact` endpoint that takes in templates & data and returns a file artifact.
 * Added `RemoveFile` to both `FileCache` and `FilesArtifactStore` to support removing files from the cache and store.